move packages to cmd, update instructions
    fixes #42
